requirements.txt
streamlit==1.31.0
google-generativeai==0.7.2
textblob
dotenv